% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Members.R
\name{getAllMembersByActivity}
\alias{getAllMembersByActivity}
\title{Get All Club Members By Activity}
\source{
chess.com public API
}
\usage{
getAllMembersByActivity(club_id, access_token = NA)
}
\arguments{
\item{club_id}{ID of the club you want the list of members for}

\item{access_token}{The access token for chess.com APIs obtained through authorization}
}
\value{
A list of members grouped by activity level (weekly, monthly,  all-time (inactive))
}
\description{
Fetches all club members sorted by activity level (weekly, monthly, all-time (inactive) for the given club
}
\seealso{
[getAllClubMembers] which returns the same data already merged into one table
}
