% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChessComClubManagement.R
\name{getAllMemberStats}
\alias{getAllMemberStats}
\title{Get Stats for all Users in the Given Club}
\source{
chess.com public API
}
\usage{
getAllMemberStats(club_id)
}
\arguments{
\item{user_id}{ID of the user you want stats for}
}
\value{
Tibble of relevant user stats for club management:
Username, joined chess.com date, last online date, country, daily standard and 960 ratings, time per move, and timeout percent
}
\description{
Fetches relevant stats for all users in the given club.
}
