% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChessComClubManagement.R
\name{getInactiveMatchPlayers}
\alias{getInactiveMatchPlayers}
\title{Find players who have not played a match in the past n days}
\source{
chess.com public API
}
\usage{
getInactiveMatchPlayers(club_id, nDays = 90)
}
\arguments{
\item{club_id}{ID of the club you want the list of inactive members for}

\item{nDays}{The number of days the user has been inactive}
}
\value{
Tibble of members who have not joined a match in the past n days along with the date they joined the club
}
\description{
Calculates which members have not joined a team match in the past 90 days
}
