% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Members.R
\name{getUserStats}
\alias{getUserStats}
\title{Get Stats for a Given User}
\source{
chess.com public API
}
\usage{
getUserStats(user_id, access_token = NA)
}
\arguments{
\item{user_id}{ID of the user you want stats for}

\item{access_token}{The access token for chess.com APIs obtained through authorization}
}
\value{
One row tibble of relevant user stats for club management: Username, joined chess.com date, last online date, country, daily standard and 960 ratings, time per move, and timeout percent
}
\description{
Returns relevant stats for a user
}
