% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Invites.R
\name{getUsersToInvite}
\alias{getUsersToInvite}
\title{Get a list of users to invite to your club}
\source{
Chess.com public API
}
\usage{
getUsersToInvite(
  club_id,
  max_timeout = NA,
  max_move_speed = NA,
  min_games = 0,
  min_months_account_age = 0,
  min_days_last_online = NA,
  min_rating = 0,
  country_code = NA,
  access_token = NA
)
}
\arguments{
\item{club_id}{ID of the club you want to invite members from}

\item{max_timeout}{The maximum allowed timeout percentage required to participate}

\item{max_move_speed}{The maximum time per move required to be invited (in hours)}

\item{min_games}{The minimum number of games that must be completed before joining}

\item{min_months_account_age}{The minimum age of the user's account}

\item{min_days_last_online}{The number of days ago since the user was last online}

\item{min_rating}{The minimum allowed rating of the user}

\item{country_code}{The country code on the user's profile. Each country has a 2 character code assigned by chess.com}

\item{access_token}{The access token for chess.com APIs obtained through authorization}
}
\value{
Data frame of ideal players within a given club
}
\description{
Get a list of ideal players within a given club based on provided parameters
}
