% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChessComClubManagement.R
\name{convertCountryCode}
\alias{convertCountryCode}
\title{Convert the Country Code from User Stats}
\source{
chess.com public API
}
\usage{
convertCountryCode(countryEndpoint)
}
\arguments{
\item{countryEndpoint}{the URL for the chess.com country API}
}
\value{
The name of the country
}
\description{
Calls the chess.com country API to get the country name. Provided only for convenience since all returns by default will not convert the country.
}
\examples{
member_data \%>\%
 mutate(
  country = sapply(country, convertCountryCode)
 )
}
