% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Members.R
\name{getAllClubMembers}
\alias{getAllClubMembers}
\title{Get All Club Members}
\source{
chess.com public API
}
\usage{
getAllClubMembers(club_id, access_token = NA)
}
\arguments{
\item{club_id}{ID of the club you want the members of}

\item{access_token}{The access token for chess.com APIs obtained through authorization}
}
\value{
A Tibble of all members in the club and their join date
}
\description{
Fetches all club members for the given club and returns the usernames and dates they joined the club.
}
