% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Games.R
\name{getAllGamesForPlayer}
\alias{getAllGamesForPlayer}
\title{Get the stats on the timeouts for a player}
\source{
chess.com public API
}
\usage{
getAllGamesForPlayer(user_id, year, month, nmonths, access_token = NA)
}
\arguments{
\item{user_id}{ID of the user you want stats for}

\item{year}{Which year to start querying the user's archive from}

\item{month}{Which month to start querying the user's archive from}

\item{nmonths}{The number of months to query the archives}

\item{access_token}{The access token for chess.com APIs obtained through authorization}
}
\value{
All daily games for a player
}
\description{
Fetches a player's daily games from the archive
}
\seealso{
[getGameResultsForPlayer()] `getGameResultsForPlayer`
}
