\name{getAllTimeLeaderboard}
\alias{getAllTimeLeaderboard}
\title{Get All Time Leaderboard for Team Matches}
\usage{
getAllTimeLeaderboard(clubId)
}
\description{
Fetches the record of match results for all players who have ever played a match for the club
}
\examples{
getAllTimeLeaderboard(clubId = "my-club")
}
