% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Matches.R
\name{getAllTimeLeaderBoard}
\alias{getAllTimeLeaderBoard}
\title{Get All Time Leaderboard for Team Matches}
\source{
chess.com public API
}
\usage{
getAllTimeLeaderBoard(club_id)
}
\arguments{
\item{club_id}{ID of the club you want the leader board for}
}
\value{
A Tibble of club members who have participated in matches and their all-time records
}
\description{
Fetches the record of match results for all players who have ever played a match for the club and are currently members of the club.
Note: chess.com is temporarily limiting finished matches to the most recent 500 due to performance.
}
\note{
Chess.com public API only returns the most recent 500 completed matches due to performance issues. Until that is resolved by the chess.com team, this function may NOT return the all-time stats and results may not be accurate for all clubs.
}
