% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChessComClubManagement.R
\name{getGameResultsForPlayer}
\alias{getGameResultsForPlayer}
\title{Get the results and the timeouts of daily games for a player}
\source{
chess.com public API
}
\usage{
getGameResultsForPlayer(
  user_id,
  year,
  month,
  nmonths,
  include_vacation = FALSE
)
}
\arguments{
\item{user_id}{ID of the user you want stats for}

\item{year}{Which year to start querying the user's archive from}

\item{month}{Which month to start querying the user's archive from}

\item{nmonths}{The number of months to query the archives}

\item{include_vacation}{Boolean for whether to check the player's tournament games for vacation rules}
}
\value{
Tibble of the number of games for a user for each event type/time control/result grouping
}
\description{
Query a player's game archive to find all match and tournament games
}
