% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Matches.R
\name{getMatchDetailsForMatches}
\alias{getMatchDetailsForMatches}
\title{Get details for club matches}
\source{
chess.com public API
}
\usage{
getMatchDetailsForMatches(club_id, match_ids, access_token = NA)
}
\arguments{
\item{club_id}{ID of the club you want matches details for}

\item{match_ids}{List of IDs of the matches you want details for}

\item{access_token}{The access token for chess.com APIs obtained through authorization}
}
\value{
Tibble of all specified match data
}
\description{
Fetches match data for the given list of team matches for a particular club Id.
}
