\name{getMatchDetailsForMatches}
\alias{getMatchDetailsForMatches}
\title{Get Details for Club Matches}
\usage{
getMatchDetailsForMatches(clubId, match_ids)
}
\description{
Fetches match data for the given list of team matches for a particular club Id.
}
\examples{
getMatchDetailsForMatches(clubId = "my-club", match_ids = c(100, 101, 102))
}
