% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Matches.R
\name{getMatchIds}
\alias{getMatchIds}
\title{Get the Daily Match Ids for a given club}
\source{
chess.com public API
}
\usage{
getMatchIds(
  club_id,
  include_finished = TRUE,
  include_in_progress = TRUE,
  include_upcoming = TRUE,
  nDays = NA,
  access_token = NA
)
}
\arguments{
\item{club_id}{ID of the club you want match IDs for}

\item{include_finished}{Returns finished match IDs}

\item{include_in_progress}{Returns in-progress match IDs}

\item{include_upcoming}{Returns upcoming match IDs}

\item{nDays}{The number of days to look back for matches, based on the match start date}

\item{access_token}{The access token for chess.com APIs obtained through authorization}
}
\value{
List of match IDs
}
\description{
Fetches the daily match Ids for the given club and returns a list of Ids.
You can specify which match status to include (finished, in-progress, and upcoming) as well as the number of days in the past the match started.
A value of nDays = NA will return all matches that are available from the chess.com API.
}
