\name{getMatchResults}
\alias{getMatchResults}
\title{Get the results of team matches against other clubs}
\usage{
getMatchResults(clubId)
}
\description{
Fetches the completed team matches and returns a the given club's record against other clubs.
}
\examples{
getMatchResults(clubId = "my-club")
}
