% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChessComClubManagement.R
\name{getMatchResults}
\alias{getMatchResults}
\title{Get the results of team matches against other clubs}
\source{
chess.com public API
}
\usage{
getMatchResults(club_id)
}
\arguments{
\item{club_id}{ID of the club you want the results for}
}
\value{
A Tibble of the club's team match record versus other clubs
}
\description{
Fetches the completed team matches and returns a the given club's record against other clubs.
}
\note{
Chess.com public API only returns the most recent 500 completed matches due to performance issues. Until that is resolved by the chess.com team, this function may NOT return all the clubs ever played.
}
