% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Matches.R
\name{getMatchUrls}
\alias{getMatchUrls}
\title{Get the API URLs of matches for a given club}
\source{
chess.com public API
}
\usage{
getMatchUrls(
  club_id,
  include_finished = TRUE,
  include_in_progress = TRUE,
  include_upcoming = TRUE,
  nDays = NA
)
}
\arguments{
\item{club_id}{ID of the club you want match URLs for}

\item{include_finished}{Returns finished match URLs}

\item{include_in_progress}{Returns in-progress match URLs}

\item{include_upcoming}{Returns upcoming match URLs}

\item{nDays}{The number of days to look back for matches, based on the match start date}
}
\value{
List of match URLs
}
\description{
Fetches the URLs for the given club and returns a list of URLs.
You can specify which match status to include (finished, in-progress, and upcoming) as well as the number of days in the past the match started.
A value of nDays = NA will return all match URLs that are available from the chess.com API.
}
\seealso{
[getMatchIds()]
}
