% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Matches.R
\name{getPlayersToRemoveFromMatch}
\alias{getPlayersToRemoveFromMatch}
\title{Get a list of users ineligible to participate in a given match}
\source{
chess.com public API
}
\usage{
getPlayersToRemoveFromMatch(match_id, club_id, max_timeouts, min_total_games)
}
\arguments{
\item{match_id}{ID of the match to manage}

\item{club_id}{ID of the club you manage}

\item{max_timeouts}{The maximum timeout percentage allowed for participation}

\item{min_total_games}{The minimum number of completed games allowed for participation}
}
\value{
A list of players ineligible to participate in the match, filtered by the provided criteria
}
\description{
Retrieves players registered for a match that are ineligible to participate
}
